+++ ats-mini/Common.h
@@ -238,6 +238,7 @@ bool identifyFrequency(uint16_t freq, bool periodic = false);
 // Network.cpp
 int8_t getWiFiStatus();
 char *getWiFiIPAddress();
+int getWiFiRSSI();
 char *getOTAIPAddress();
 void netInit(uint8_t netMode, bool showStatus = true);
 void netInitServices();